Design the app with bulma  - lot of tiny edits  - unused files removed  - some js files renamed to jsx
