--- src/components/App/styles.css
@@ -1,4 +1,0 @@
-* {
-  margin: 0px;
-  padding: 0px;
-}
