--- src/components/Footer/Footer.js
@@ -1,26 +1,0 @@
-import React from 'react';
-
-const Footer = () => (
-    <footer className="footer">
-        <div className="content has-text-centered">
-            <div className="footer__icons-notice">
-                <p>
-                    Icons made by{' '}
-                    <a
-                        href="https://www.flaticon.com/authors/smashicons"
-                        title="Smashicons"
-                    >
-                        Smashicons
-                    </a>{' '}
-                    from{' '}
-                    <a href="https://www.flaticon.com/" title="Flaticon">
-                        www.flaticon.com
-                    </a>
-                </p>
-            </div>
-            <div>(c) 2019 by nopefish</div>
-        </div>
-    </footer>
-);
-
-export { Footer };
