+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Footer = () => (
+    <footer className="footer">
+        <div className="content has-text-centered">
+            <div>(c) 2019 by nopefish</div>
+        </div>
+    </footer>
+);
+
+export { Footer };
