--- src/components/Home/Home.css
@@ -1,4 +1,0 @@
-.filters-block {
-  padding: 20px;
-  /* background-color: #a9eeea; */
-}
