+++ src/components/MovieList/MovieList.jsx
@@ -22,10 +22,10 @@ export const MovieList = ({ movies = [], history }) => {
         };
         return (
             <div
-                className="content__movie-short-info columns is-desktop tile"
+                className="movie-list__short-info columns is-desktop"
                 onClick={gotoMovieDetails}
             >
-                <div className="content__movie-short-info-poster column is-2-desktop is-full-mobile is-2-tablet">
+                <div className="movie-list__short-info-poster column is-2-desktop is-full-mobile is-2-tablet">
                     <ImgFallback
                         alt="Movie poster"
                         src={posterUrl}
@@ -33,26 +33,26 @@ export const MovieList = ({ movies = [], history }) => {
                     />
                 </div>
 
-                <div className="content__movie-short-info-text-part column">
-                    <div className="content__movie-short-info-title-year-genres">
-                        <div className="content__movie-short-info-title">
+                <div className="movie-list__short-info-text-part column">
+                    <div className="movie-list__short-info-title-year-genres">
+                        <div className="movie-list__short-info-title">
                             {title}
                         </div>
-                        <div className="content__movie-short-info-year">
+                        <div className="movie-list__short-info-year">
                             {year}
                         </div>
-                        <div className="content__movie-short-info-genres">
+                        <div className="movie-list_-short-info-genres">
                             {genres.map((g, i) => (
                                 <span key={i} className="tag is-info">{g}</span>
                             ))}
                         </div>
                     </div>
 
-                    <div className="content__movie-short-info-director">
+                    <div className="movie-list__short-info-director">
                         {director}
                     </div>
-                    <div className="content__movie-short-info-plot">{plot}</div>
-                    <div className="content__movie-short-info-runtime">
+                    <div className="movie-list__short-info-plot">{plot}</div>
+                    <div className="movie-list__short-info-runtime">
                         {runtime} min.
                     </div>
                 </div>
@@ -61,7 +61,7 @@ export const MovieList = ({ movies = [], history }) => {
     };
 
     return (
-        <div className="content__movie-list">
+        <div className="movie-list">
             {movies.map(m => (
                 <MovieShortInfo movie={m} key={m.id} />
             ))}
